calendar_bot/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
